refactor(watchdog): move the process self-watchdog out of recovery.rs (#5147)
Move-only. `spawn_watchdog` and its four timing constants leave
`health/recovery.rs` for `health/recovery/self_watchdog.rs` byte-for-byte;
`recovery.rs` re-exports the symbol, so `health::spawn_watchdog` and its single
call site in `cli/dcserver.rs` are untouched.

Two reasons, both prerequisites for the #5147 work that follows.

Cohesion. `recovery.rs` recovers *channels* — watchers, mailboxes, stalled
turns. This recovers the *process*: it runs on its own OS thread precisely so it
depends on nothing the rest of that module touches, and it shares no state with
any of it.

Reach. `CHECK_INTERVAL`, `TCP_TIMEOUT` and `MAX_FAILURES` were function-local
`const`s, so nothing outside `spawn_watchdog` could name them. The forensics
module that lands next derives its runtime-staleness threshold from
`TCP_TIMEOUT` and has to be able to check that derivation. At module scope it
can import the value; from inside a function body the only available technique
is `include_str!` plus `str::contains` on the declaration text, which sees a
file as one flat string and cannot tell code from a comment.

`audit_maintainability_giant_baseline.toml` for `recovery.rs` is LOWERED
2750 -> 2660. The file goes from 2737 to 2642 production LoC — 95 lines
recovered — and the ceiling takes 90 of them, so headroom goes 13 -> 18 rather
than to 0. That is deliberate: a ratchet pinned to a file's exact current size
fails a hard gate on the next one-line correctness fix before it can be
reviewed, on the file this investigation is most likely to touch. (2750 was the
old ceiling, not the old size; the file arrived at this branch 13 lines under
it.) No ceiling anywhere is raised by this series. Verified: `cargo check --workspace --all-targets`,
`cargo test --lib services::discord::health::recovery` (59 passed),
`audit_maintainability.py --check`, `check_test_lane_coverage.py`,
`check_test_target_integrity.py`, `check-ci-runner-hardening.sh` all rc=0.

Refs #4756, #4770
