+++ src/services/discord/health.rs
@@ -63,6 +63,11 @@ pub use headless_turn::{
     start_reserved_headless_agent_turn_with_owner_channel,
 };
 pub use mailbox::purge_idle_channel_mailbox_registry_entry;
+// #5147: re-exported as a module (not as loose constants) so a consumer reads
+// `self_watchdog::TCP_TIMEOUT` — a name that says which timeout — rather than a
+// bare `TCP_TIMEOUT` at the health root.
+#[allow(unused_imports)]
+pub(crate) use recovery::self_watchdog;
 #[allow(unused_imports)]
 pub use recovery::{
     HardStopRuntimeResult, IdleTmuxStaleTurnRepairResult, PendingQueueSnapshot,
